refactor(model): split StatisticsResponse.kt into per-type files
Prep for /status response — RecordingsStatistics and DetectServerStatistics
will be reused; StatisticsResponse wrapper will be removed later.
