+++ modules/model/src/main/kotlin/ru/zinin/frigate/analyzer/model/response/RecordingsStatistics.kt
@@ -1,0 +1,16 @@
+package ru.zinin.frigate.analyzer.model.response
+
+data class RecordingsStatistics(
+    val total: Long,
+    val processed: Long,
+    val unprocessed: Long,
+    val byCameras: List<CameraStatistics>,
+    val processingRatePerMinute: Double,
+)
+
+data class CameraStatistics(
+    val camId: String,
+    val recordingsCount: Long,
+    val recordingsProcessed: Long,
+    val detectionsCount: Long,
+)
